Created new package for tools related to Specify, but not needed by the main app.
Added DataModelWriter.java which takes an inventory of 
all of the HBM files in the datamodel directory, and generates
one single XML file that describes the the Specify datamodel.

Associated files are (java and betwixt) for Field, Table,
and Relationship.  These are stored in a subpackage.

Still need to add TableID's to this datamodel file.