Lots of changes to my beta code...
TreeDataListModel:
Fixed a couple of bugs related to tracking the visible node count per rank.
Corrected a comment.
Fix misspelling in method name.
Reworked getVisibleRanks to be more efficient.
Modified the model to always return itself as the first element to signal the renderer to display a header instead of an actual model element.
Modified getElementAt and getSize to handle those changes.
Modified getLongestNameAndPixelLengthByRank to take the rank name into account.
Modified the fireInterval(Added/Removed) calls to send the proper indices, adjusting for the first element being the model itself.

TreeDataListCellRenderer:
Modified renderer to display a simple header as the first list element.

Test:
Put the list in a scrollpane.