Lots of work on the attachment storage system.
Updated Attachment class to handle setting the attachmentLocation after the orig filename is given.
Updated Attachment class to copy the files to the attachment storage system on DB commits.

Updated BuildSampleDatabase to create some sample attachments.

Added a couple of methods to AttachmentManagerIface to handle file naming issues as well as to provide a cleanup() callback to the attachment manager at system shutdown.

Modified Specify.java to start and stop the attachment system at startup and shutdown.