Oops.  This was missing from my earlier checkin.  It shouldn't have caused a problem.
This takes care of removing the deg symbol (°) and tick marks for minutes and seconds (' and ") from input to the LatLonConverter.
It also fixed a problem with stripping zeros from a number like 0.620000.  Previously the leading 0 was also stripped.  That's pretty much never what you want.  I changed a call to StringUtils.strip() to StringUtils.stripEnd().