temporary fix for bug #4504 Null pointer occurs when re-import an exported Simple DataSet
drawback to this fix means that the system doesn't catch that there is a
dataset mapping already defined for this file (which is should have since the
data was exported from and existing dataset).  

it doesn't ask you if you want to reuse and existing mapping, it goes straight
to the mapping editor.