Added popups when the user tries to do something to the tree def that would require changes to tree nodes.
For example:
If the user edits a tree level and changes anything that would require updating of a lot of "fullname" fields, a popup is shown and the action is stopped.
If the user tries to enforce a tree level, but nodes in the tree violate that rule, a popup is shown and the action is blocked.