massive changes to the db converter code to support conversion between sql server, mysql. making many sql call sql standard. making many result set, statement, and query functions db agnostic tested on an additional 9 databases, and code bug fixes tightended up. some changes to datamodel.
following email and instructions sent out for use of new code:
I checked in all of the changes to the DB converter to support SQL Server conversions, as well as many fixes to the converter to support conversions of other databases.  

We now have 9+ databases available for converting and testing.  The mySQL dumps can be found at:
\\huxley\specify6\Specify5_Dbs_MySql

Of importance: RunMyContentsAtCommandLine.txt – 
Take the contents of this file, and run them at the commandline.  It references 2 script files “createSp5dbs.sql” and “createSp6dbs.sql” (make sure you are cd’d into this folder before running the commands).  This will get all of the necessary databases created, “attached” and permissioned on your MySQL server.  It will perform all of the necessary steps so that you can run “automatic” conversions (converter dialog will work at it did before with a picklist of available conversions for mysql - > mysql conversions).  There is also a text file that lists what Specify 5.X version the database is.

I’ve checked things in a little more prematurely than ideal, I didn’t want to hold up any longer, and the remaining fixes are fairly minor.  I can guarantee that the conversion works on kufish.  I’ve got to rerun the tests on the remaining dbs and make a couple tweaks. 

Don’t pull and expect to do a db conversion until you copy the new dumps and run the scripts.  I’ve change the code and removed all references to the old ones.

Anyone wanting to run a M$ server conversion will need to add the jtds jar to their build path.  Otherwise, it can be ignored.  Find it in the libs folder.